Adds MVVM architecture
