+++ lib/enums/view_state.dart
@@ -1,0 +1,1 @@
+enum ViewState { Idle, Busy, Error }
