+++ lib/ui/components/feature_card.dart
