+++ lib/ui/components/loading_indicator.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class LoadingIndicator extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: MediaQuery.of(context).size.height,
+      width: MediaQuery.of(context).size.width,
+      child: Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
