+++ lib/ui/components/teachers_card.dart
