+++ lib/ui/components/team_card.dart
