+++ lib/ui/views/base_view.dart
@@ -1,0 +1,34 @@
+import 'package:cv_projects_task/locator.dart';
+import 'package:cv_projects_task/viewmodels/base_model.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class BaseView<T extends BaseModel> extends StatefulWidget {
+  final Widget Function(BuildContext context, T model, Widget child) builder;
+  final Function(T) onModelReady;
+
+  BaseView({this.builder, this.onModelReady});
+
+  @override
+  _BaseViewState<T> createState() => _BaseViewState<T>();
+}
+
+class _BaseViewState<T extends BaseModel> extends State<BaseView<T>> {
+  T model = locator<T>();
+
+  @override
+  void initState() {
+    if (widget.onModelReady != null) {
+      widget.onModelReady(model);
+    }
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ChangeNotifierProvider<T>(
+      create: (context) => model,
+      child: Consumer<T>(builder: widget.builder),
+    );
+  }
+}
