+++ lib/viewmodels/base_model.dart
@@ -1,0 +1,21 @@
+import 'package:cv_projects_task/enums/view_state.dart';
+import 'package:flutter/material.dart';
+
+class BaseModel extends ChangeNotifier {
+  ViewState _state = ViewState.Idle;
+  String _errorMessage = "";
+
+  String get errorMessage => _errorMessage;
+
+  void setErrorMessage(String errorMessage) {
+    _errorMessage = errorMessage;
+    notifyListeners();
+  }
+
+  ViewState get state => _state;
+
+  void setState(ViewState viewState) {
+    _state = viewState;
+    notifyListeners();
+  }
+}
